requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.8
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
flatbuffers==24.3.25
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
humanfriendly==10.0
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
jsonpath-python==1.0.6
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mistralai==1.1.0
mpmath==1.3.0
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==2.0.2
onnxruntime==1.19.2
openai==1.44.1
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==10.4.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
python-dateutil==2.8.2
pytz==2024.2
PyYAML==6.0
rapid-layout==0.3.0
requests==2.32.3
rich==13.9.2
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.2
tenacity==9.0.0
thinc==8.3.2
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
scipy==1.11.3
scikit-learn==1.3.1
